#75 Implement functionality to draw a path
Added a way to add 2 points manually on the map, and then a line will be automatically drawn between the 2 points. The camera will zoom on the first point(green point). More than 2 points can't be added, if an attempt to place 3 markers is made, the 2 past markers are deleted.